xponencIA-edicion-main/xponencIA-edicion-main/requirements.txt
openai-whisper
moviepy
scenedetect[opencv]
google-generativeai
python-dotenv